# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Plant API",
        "description": "Ugne's Simple Plant API",
        "contact": {
            "email": "n00201661@student.iadt.ie"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/climates?page={page}": {
            "get": {
                "tags": [
                    "Climates"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Displays all climates, page by page",
                "operationId": "33702dbd55ac840a22bac5b5d7d8b454",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, Returns a list of Categories in JSON format"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/climates": {
            "post": {
                "tags": [
                    "Climates"
                ],
                "summary": "Create a new Climate",
                "description": "Stores the climate in the database",
                "operationId": "storeClimate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A quick description about this climate"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/climates/{id}": {
            "get": {
                "tags": [
                    "Climates"
                ],
                "summary": "Display the specified resource.",
                "description": "Gets a climate by id",
                "operationId": "7987fadc8afc3a41c3d064f096ccf5d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Climate id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Climates"
                ],
                "summary": "Update a climate",
                "description": "Updates the climate in the database",
                "operationId": "putClimate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Climate id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A quick description about this climate"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Climates"
                ],
                "summary": "Delete a Climate",
                "description": "Delete Climate",
                "operationId": "destroyClimate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the Climate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "204"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}